Add experimental deterministic job feature
This change inspects jobs before enquing and performs them immediately
if the process is already running in the background.

I.E.
If a foreground process (Brwoser UI) submits a job, it's added to
the queuing system the same as always.

If a process already running in a background worker submits a job,
execute it immediately in the same thread.

**RATIONALE**
En-queuing and De-queuing jobs reuqires them to be serialized,
saved to the database, then deserialized, and usually arguments
reloaded from the database to ensure current state. We want to
determine how much actual time the serialization/deserialization
and object reloading adds to an overall import process.